move NotEmptyTransformation to its own file
