+++ include/ensmallen_bits/cmaes/active_cmaes.hpp
@@ -116,7 +116,7 @@ class ActiveCMAES
                        const size_t maxIterations = 1000,
                        const double tolerance = 1e-5,
                        const SelectionPolicyType& selectionPolicy = SelectionPolicyType(),
-                       double stepSize = 1);
+                       double stepSize = 0);
 
   /**
    * Optimize the given function using Active CMA-ES. The given starting point will be
