+++ include/ensmallen_bits/cmaes/active_cmaes_impl.hpp
@@ -12,39 +12,14 @@
  * the 3-clause BSD license along with ensmallen.  If not, see
  * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
-#ifndef ENSMALLEN_ACTIVE_CMAES_CMAES_IMPL_HPP
-#define ENSMALLEN_ACTIVE_CMAES_CMAES_IMPL_HPP
+#ifndef ENSMALLEN_CMAES_ACTIVE_CMAES_IMPL_HPP
+#define ENSMALLEN_CMAES_ACTIVE_CMAES_IMPL_HPP
 
 // In case it hasn't been included yet.
 #include "active_cmaes.hpp"
 
+#include "not_empty_transformation.hpp"
 #include <ensmallen_bits/function.hpp>
-
-#ifndef NOT_EMPTY_TRANSFORMATION
-#define NOT_EMPTY_TRANSFORMATION
-
-/*
- This partial specialization is used to throw an exception when the
- TransformationPolicyType is EmptyTransformation and call a
- constructor with parameters 'lowerBound' and 'upperBound' otherwise.
- This shall be removed when the deprecated constructor is removed in
- the next major version of ensmallen.
-*/
-template<typename T1, typename T2>
-struct NotEmptyTransformation : std::true_type {
-  void Assign(T1& obj, double lowerBound, double upperBound) {
-    obj = T1(lowerBound, upperBound);
-  }
-};
-
-template<template<typename...> class T, typename... A, typename... B>
-struct NotEmptyTransformation<T<A...>, T<B...>> : std::false_type {
-  void Assign(T<A...>& obj, double lowerBound, double upperBound) {
-    throw std::logic_error("TransformationPolicyType is EmptyTransformation");
-  }
-};
-
-#endif
 
 namespace ens {
